+++ kernel/SlippiMemory.h
@@ -37,7 +37,7 @@ typedef struct SlpReadResult
 typedef struct SlpGameReader
 {
 	u16 payloadSizes[PAYLOAD_SIZES_BUFFER_SIZE];
-	u32 lastReadPos;
+	u64 lastReadPos;
 	SlpMetadata metadata;
 	SlpReadResult lastReadResult;
 } SlpGameReader;
@@ -48,6 +48,6 @@ void SlippiMemoryWrite(const u8 *buf, u32 len);
 
 // The same game object should be passed to this function every call. This function
 // will always read full payloads into the buffer, partial payloads will not be loaded
-SlpMemError SlippiMemoryRead(SlpGameReader *reader, u8 *buf, u32 bufLen, u32 readPos);
+SlpMemError SlippiMemoryRead(SlpGameReader *reader, u8 *buf, u32 bufLen, u64 readPos);
 
 #endif
